PoC: native RHP stacks in JS
