--- src/libs/Navigation/PlatformStackNavigation/navigationOptions/convertToWebNavigationOptions.ts
@@ -1,24 +1,0 @@
-import type {PlatformStackNavigationOptions, PlatformStackScreenProps} from '@libs/Navigation/PlatformStackNavigation/types';
-import {isRouteBasedScreenOptions} from '@libs/Navigation/PlatformStackNavigation/types';
-
-import type {ParamListBase, ScreenOptionsOrCallback} from '@react-navigation/native';
-import type {StackNavigationOptions} from '@react-navigation/stack';
-
-import buildPlatformSpecificNavigationOptions from './buildPlatformSpecificNavigationOptions';
-
-function convertToWebNavigationOptions(screenOptions: ScreenOptionsOrCallback<PlatformStackNavigationOptions> | undefined): ScreenOptionsOrCallback<StackNavigationOptions> | undefined {
-    if (!screenOptions) {
-        return undefined;
-    }
-
-    if (isRouteBasedScreenOptions(screenOptions)) {
-        return (props: PlatformStackScreenProps<ParamListBase, string>) => {
-            const routeBasedScreenOptions = screenOptions(props);
-            return {...buildPlatformSpecificNavigationOptions(routeBasedScreenOptions), ...routeBasedScreenOptions.web};
-        };
-    }
-
-    return {...buildPlatformSpecificNavigationOptions<StackNavigationOptions>(screenOptions), ...screenOptions.web};
-}
-
-export default convertToWebNavigationOptions;
